Move page, fix grammatical error
