+++ apps/launch/components/NominateContract/styles.tsx
